Modular Refactor
This is a big, but long overdue refactor to move away from a single unwieldy file to a much more modular and readable setup. These changes obviously mean that config scripts will need to be refactored, but I've tried to keep this minimal to the outward-facing user, and hopefully this makes things more simple going forwards.

For now, the old scripts are kept in the LEGACY directory

Notable changes:

* Overwrites are now defined in the config per-step file, rather than in a big list at the top
* Parameters are now combined into a single big dictionary, which means in practise one more level at the config.toml stage. This means adding steps in the future will make things significantly easier
* Download/wcs_adjust/release now happens within the main pipeline itself, rather than in separate scripts
* destripe is renamed to single_tile_destripe
* dither_stripe_sub is renamed to multi_tile_destripe
* dither_match is now level_match
* A new step has been added, lyot_separate, which should allow us to include the MIRI coronagraph in the mosaics by splitting out the coronagraph and main science chips and treating separately